refactor: move accounts and neurons modal components to components
