+++ frontend/svelte/src/lib/components/neurons/ConfirmDissolveDelay.svelte
@@ -1,7 +1,7 @@
 <script lang="ts">
   import type { NeuronInfo } from "@dfinity/nns";
   import { createEventDispatcher } from "svelte";
-  import Spinner from "../../components/ui/Spinner.svelte";
+  import Spinner from "../ui/Spinner.svelte";
   import { updateDelay } from "../../services/neurons.services";
   import { i18n } from "../../stores/i18n";
   import { secondsToDuration } from "../../utils/date.utils";
